feature/bugfix: AuthControllerTests.cs file got modified, PublicWebApplication.cs moved to a global class.
